update int to long long

+++ src/coros/memory/buffer.cpp
@@ -5,7 +5,7 @@
 #include <cstring>
 #include <stdexcept>
 
-coros::base::ByteBuffer::ByteBuffer(int max_capacity) {
+coros::base::ByteBuffer::ByteBuffer(long long max_capacity) {
     this->data = new std::byte[max_capacity];
     this->read_p = 0;
     this->write_p = 0;
@@ -16,18 +16,18 @@ coros::base::ByteBuffer::~ByteBuffer() {
     delete[] this->data;
 }
 
-int coros::base::ByteBuffer::get_index(int p) {
+long long coros::base::ByteBuffer::get_index(long long p) {
     return p % max_capacity;
 }
 
 bool coros::base::ByteBuffer::has_wrap_around() {
-    int quotient_w = write_p / max_capacity;
-    int quotient_r = read_p / max_capacity;
+    long long quotient_w = write_p / max_capacity;
+    long long quotient_r = read_p / max_capacity;
     return quotient_w > quotient_r && read_p > (quotient_r * max_capacity);
 }
 
 coros::base::IOChunk coros::base::ByteBuffer::get_read_chunk() {
-    int read_index = get_index(read_p);
+    long long read_index = get_index(read_p);
     if (!has_wrap_around()) {
         return { data + read_index, write_p - read_p };
     }
@@ -35,14 +35,14 @@ coros::base::IOChunk coros::base::ByteBuffer::get_read_chunk() {
 }
 
 coros::base::IOChunk coros::base::ByteBuffer::get_write_chunk() {
-    int write_index = get_index(write_p);
+    long long write_index = get_index(write_p);
     if (has_wrap_around()) {
         return { data + write_index, get_index(read_p) - write_index };
     }
     return { data + write_index, max_capacity - write_index };
 }
 
-void coros::base::ByteBuffer::increment_read_pointer(int size) {
+void coros::base::ByteBuffer::increment_read_pointer(long long size) {
     if (size > get_total_remaining()) {
         throw std::runtime_error("ByteBuffer increment_read_pointer error: More than remaining");
     }
@@ -53,20 +53,20 @@ void coros::base::ByteBuffer::increment_read_pointer(int size) {
     }
 }
 
-void coros::base::ByteBuffer::increment_write_pointer(int size) {
+void coros::base::ByteBuffer::increment_write_pointer(long long size) {
     if (size > get_total_capacity()) {
         throw std::runtime_error("ByteBuffer increment_write_pointer error: More than capacity");
     }
     write_p += size;
 }
 
-void coros::base::ByteBuffer::read(std::byte* dest, int size) {
+void coros::base::ByteBuffer::read(std::byte* dest, long long size) {
     if (size > get_total_remaining()) {
         throw std::runtime_error("ByteBuffer read error: Read size more than remaining");
     }
     while (size > 0) {
         IOChunk chunk = get_read_chunk();
-        int read_size = std::min(size, chunk.size);
+        long long read_size = std::min(size, chunk.size);
         std::memcpy(dest, chunk.data, read_size);
         size -= read_size;
         increment_read_pointer(read_size);
@@ -82,13 +82,13 @@ std::byte coros::base::ByteBuffer::read_b() {
     return b;
 }
 
-void coros::base::ByteBuffer::write(std::byte* src, int size) {
+void coros::base::ByteBuffer::write(std::byte* src, long long size) {
     if (size > get_total_capacity()) {
         throw std::runtime_error("ByteBuffer write error: Write size more than capacity");
     }
     while (size > 0) {
         IOChunk chunk = get_write_chunk();
-        int write_size = std::min(size, chunk.size);
+        long long write_size = std::min(size, chunk.size);
         std::memcpy(chunk.data, src, write_size);
         size -= write_size;
         increment_write_pointer(write_size);
@@ -103,10 +103,10 @@ void coros::base::ByteBuffer::write_b(std::byte b) {
     increment_write_pointer(1);
 }
 
-int coros::base::ByteBuffer::get_total_capacity() {
+long long coros::base::ByteBuffer::get_total_capacity() {
     return max_capacity - get_total_remaining();
 }
 
-int coros::base::ByteBuffer::get_total_remaining() {
+long long coros::base::ByteBuffer::get_total_remaining() {
     return write_p - read_p; 
 }
