+++ src/coros/memory/buffer.h
@@ -6,30 +6,30 @@
 namespace coros::base {
     struct IOChunk {
         std::byte* data;
-        int size;
+        long long size;
     };
 
     class ByteBuffer {
         private:
             std::byte* data;
-            int read_p;
-            int write_p;
-            int max_capacity;
-            int get_index(int p);
+            long long read_p;
+            long long write_p;
+            long long max_capacity;
+            long long get_index(long long p);
             bool has_wrap_around();
         public:
-            ByteBuffer(int max_capacity);
+            ByteBuffer(long long max_capacity);
             ~ByteBuffer();
             IOChunk get_read_chunk();
             IOChunk get_write_chunk();
-            void increment_read_pointer(int size);
-            void increment_write_pointer(int size);
-            void read(std::byte* dest, int size);
+            void increment_read_pointer(long long size);
+            void increment_write_pointer(long long size);
+            void read(std::byte* dest, long long size);
             std::byte read_b();
-            void write(std::byte* src, int size);
+            void write(std::byte* src, long long size);
             void write_b(std::byte b);
-            int get_total_capacity();
-            int get_total_remaining();
+            long long get_total_capacity();
+            long long get_total_remaining();
     };
 }
